[#3116] Refactor middleware module.
It was starting to become unmanageably large with a lot of separate
concerns in one place.

- Separate the flask and pylons app code into separate modules.
- Separate the common middleware code into a separate module.
