+++ ckan/tests/config/test_middleware.py
@@ -8,7 +8,8 @@ from routes import url_for
 import ckan.plugins as p
 import ckan.tests.helpers as helpers
 
-from ckan.config.middleware import AskAppDispatcherMiddleware, CKANFlask
+from ckan.config.middleware import AskAppDispatcherMiddleware
+from ckan.config.middleware.flask_app import CKANFlask
 from ckan.controllers.partyline import PartylineController
